move selectors and type to gh helper file
